requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
Pillow==10.2.0  # For image processing
django-crispy-forms==2.1  # For better form rendering
python-magic==0.4.27  # For file type validation
django-cleanup==8.1.0  # For automatic file cleanup
whitenoise==6.6.0  # For serving static files
gunicorn==21.2.0  # For production deployment